pyproject.toml
[project]
name = "cool-routes"
version = "0.1.0"
description = ""
authors = [
    {name = "Cristopher Jesús Lincoleo Oyarce",email = "cris.lincoleo@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "geopandas (>=1.1.2,<2.0.0)",
    "folium (>=0.20.0,<0.21.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=1.24,<2.4)",
    "scipy (>=1.16.3,<2.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "rasterio (>=1.4.4,<2.0.0)",
    "osmnx (>=2.0.7,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "statsmodels (>=0.14.6,<0.15.0)",
    "numba (>=0.63.1,<0.64.0)",
    "pvlib (>=0.13.1,<0.14.0)",
    "pysolar (>=0.13,<0.14)",
    "ephem (>=4.2,<5.0)",
    "earthengine-api (>=1.7.4,<2.0.0)",
    "google-cloud-storage (>=3.7.0,<4.0.0)",
    "google-api-python-client (>=2.187.0,<3.0.0)",
    "pydrive2 (>=1.21.3,<2.0.0)",
    "google-auth-oauthlib (>=1.2.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)"
]
